Metadata

Public

Metadata.attach_metadata — Function

attach_metadata(x, metadata)

Generic method for attaching metadata to x.

Metadata.has_metadata — Function

has_metadata(x[, k; dim]) -> Bool

Returns true if x has metadata. If k is specified then checks for the existence of a metadata paired to k. If dim is specified then this checks the metadata at the corresponding dimension.

Metadata.metadata — Function

metadata(x[, k; dim])

Returns metadata from x. If k is specified then the metadata value paired to k is returned. If dim is specified then the operation is performed for metadata specific to dimension dim.

Metadata.metadata! — Function

metadata!(x, k, val[; dim])

Set x's metadata paired to k to val. If dim is specified then the metadata corresponding to that dimension is mutated.

Metadata.copy_metadata — Function

copy_metadata(src, dst) -> attach_metadata(dst, copy(metadata(src)))

Copies the the metadata from src and attaches it to dst. Note that this method specifically calls deepcopy on the metadata of src to ensure that changing the metadata of dst does not affect the metadata of src.

See also: share_metadata.

Metadata.drop_metadata — Function

drop_metadata(x)

Returns x without metadata attached.

Metadata.share_metadata — Function

share_metadata(src, dst) -> attach_metadata(dst, metadata(src))

Shares the metadata from src by attaching it to dst. The returned instance will have properties that are synchronized with src (i.e. modifying one's metadata will effect the other's metadata).

See also: copy_metadata.

Internal

Metadata.NoMetadata — Type

NoMetadata

Internal type for the Metadata package that indicates the absence of any metadata. DO NOT store metadata with the value NoMetadata().

Metadata.MetadataPropagation — Type

MetadataPropagation(::Type{T}) -> Union{Drop,Copy,Share}

When metadata of type T is attached to something should the same in memory instance be attached or a deep copy of the metadata?

Metadata.metadata_summary — Function

metadata_summary(x; left_pad::Int=0, l1=lpad(`•`, 3), l2=lpad('-', 5))

Creates summary readout of metadata for x.

Metadata.MetaArray — Type

MetaArray(parent::AbstractArray, metadata)

Custom AbstractArray object to store an AbstractArray parent as well as some metadata.

Examples

julia> using Metadata

julia> Metadata.MetaArray(ones(2,2), metadata=(m1 =1, m2=[1, 2]))
2×2 attach_metadata(::Array{Float64,2}, ::NamedTuple{(:m1, :m2),Tuple{Int64,Array{Int64,1}}})
  • metadata:
    - m1 = 1
    - m2 = [1, 2]
 1.0  1.0
 1.0  1.0

Metadata.MetaRange — Type

MetaRange(x::AbstractRange, meta)

Type for storing metadata alongside a range.

Examples

julia> using Metadata

julia> Metadata.MetaRange(1:1:2, (m1 =1, m2=[1, 2]))
attach_metadata(1:1:2, ::NamedTuple{(:m1, :m2),Tuple{Int64,Array{Int64,1}}})
  • metadata:
    - m1 = 1
    - m2 = [1, 2]

Metadata.MetaUnitRange — Type

MetaUnitRange(x::AbstractUnitRange, meta)

Type for storing metadata alongside a anything that is subtype of AbstractUnitRange.

Examples

julia> using Metadata

julia> Metadata.MetaUnitRange(1:2, (m1 =1, m2=[1, 2]))
attach_metadata(1:2, ::NamedTuple{(:m1, :m2),Tuple{Int64,Array{Int64,1}}})
  • metadata:
    - m1 = 1
    - m2 = [1, 2]